rxpulse/requirements.txt
praw==7.8.1
python-dotenv==1.0.1
boto3==1.35.70
pandas==2.2.3
pyarrow==18.1.0
pyspark==3.5.3
streamlit==1.40.2
plotly==5.24.1
google-generativeai==0.8.3
tenacity==9.0.0